warning
git-svn-id: https://subversion.assembla.com/svn/fakesmc@201 958d8502-8caa-4642-8a4d-47b05a98fa7a

+++ Plug-ins/FakeSMCACPImonitor/FakeSMCACPImonitor.cpp
@@ -129,7 +129,7 @@ bool ACPImonitor::start(IOService * provider)
 		{
 			OSString* name = OSDynamicCast(OSString, fanIDs->getObject(i)); 
 			FanName = name->getCStringNoCopy();
-			snprintf(key, 5, "F%dID", FanOffset + i);
+			snprintf(key, 5, "F%dID", (int)(FanOffset + i));
 			FakeSMCAddKey(key, "ch8*", strlen(FanName), (char*)FanName);
 		}		
 		fanIDs->release();
